created GeneralUtilities function
